# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API",
        "description": "RESTful API for Job Portal - manage jobs, applications, and users",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/public/jobs": {
            "get": {
                "tags": [
                    "Jobs (Public)"
                ],
                "summary": "Get all job listings (Public)",
                "description": "Retrieve paginated job listings with optional search and filters",
                "operationId": "37b7b1eb7cc2d4f3c2f62c47c1f6473e",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search keyword (searches in title, company, location)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "developer"
                        }
                    },
                    {
                        "name": "company",
                        "in": "query",
                        "description": "Filter by company name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Google"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter by location",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Jakarta"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of jobs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Software Engineer"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "We are looking for...\\"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Jakarta"
                                                    },
                                                    "company": {
                                                        "type": "string",
                                                        "example": "Tech Corp"
                                                    },
                                                    "salary": {
                                                        "type": "integer",
                                                        "example": 10000000
                                                    },
                                                    "logo": {
                                                        "type": "string",
                                                        "example": "logos/company.jpg"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-11-25T10:00:00.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/public/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs (Public)"
                ],
                "summary": "Get job details",
                "description": "Retrieve single job details by ID",
                "operationId": "4905cb878352b0fc116252c79e1d98ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Software Engineer"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Job description..."
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Jakarta"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "Tech Corp"
                                        },
                                        "salary": {
                                            "type": "integer",
                                            "example": 10000000
                                        },
                                        "logo": {
                                            "type": "string",
                                            "example": "logos/company.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            }
        },
        "/api/jobs": {
            "post": {
                "tags": [
                    "Jobs (Admin)"
                ],
                "summary": "Create new job (Admin only)",
                "description": "Create a new job vacancy",
                "operationId": "46db842b9a0994748f28e0cc492a5b45",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "location",
                                    "company"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Software Engineer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "We are looking for..."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Tech Corp"
                                    },
                                    "salary": {
                                        "type": "integer",
                                        "example": 10000000
                                    },
                                    "logo": {
                                        "type": "string",
                                        "example": "logos/logo.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Created"
                                        },
                                        "job": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{id}": {
            "put": {
                "tags": [
                    "Jobs (Admin)"
                ],
                "summary": "Update job (Admin only)",
                "description": "Update existing job details",
                "operationId": "09003eadd3b848807fdd9deb993a68ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior Software Engineer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Bandung"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Tech Corp"
                                    },
                                    "salary": {
                                        "type": "integer",
                                        "example": 15000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Updated"
                                        },
                                        "job": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jobs (Admin)"
                ],
                "summary": "Delete job (Admin only)",
                "description": "Delete a job vacancy",
                "operationId": "10aafac8e299b54e3bdd7bf7d651b26b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}